+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
-## 1.0.0-b09
+## 1.0.0-b10
 
-* Initial version
+* Simple single-collection "message queue" implemented with MongoDB's find-and-modify operations
+* Lease-based locking of received messages
+* Configurable locking times
+* Proper duplicate message ID exception in the form of `UniqueMessageIdViolationException`